feat: move all cross-check components to modules/CrossCheck
